Move test utilities to test:compile
