+++ tests/build.sbt
@@ -5,8 +5,8 @@ name := "pureconfig-tests"
 crossScalaVersions ~= { _ :+ "2.13.0" }
 
 libraryDependencies ++= Seq(
-  scalaTest,
-  scalaCheck.value,
-  scalaCheckShapeless.value)
+  scalaTest.withConfigurations(Some("compile")),
+  scalaCheck.value.withConfigurations(Some("compile")),
+  scalaCheckShapeless.value.withConfigurations(Some("compile")))
 
 skip in publish := true
