+++ tests/src/main/scala/pureconfig/BaseSuite.scala
