+++ tests/src/main/scala/pureconfig/ConfigConvertChecks.scala
