+++ tests/src/main/scala/pureconfig/arbitrary/package.scala
@@ -9,8 +9,6 @@ package object arbitrary {
   implicit val arbJavaDuration = Arbitrary(genJavaDuration)
   implicit val arbFiniteDuration = Arbitrary(genFiniteDuration)
   implicit val arbInstant = Arbitrary(genInstant)
-  implicit val arbZoneId = Arbitrary(genZoneId)
-  implicit val arbZoneOffset = Arbitrary(genZoneOffset)
   implicit val arbPeriod = Arbitrary(genPeriod)
   implicit val arbYear = Arbitrary(genYear)
   implicit val arbUUID = Arbitrary(Gen.uuid)
@@ -19,4 +17,15 @@ package object arbitrary {
   implicit val arbPercentage = Arbitrary(genPercentage)
   implicit val arbJavaBigDecimal = Arbitrary(genJavaBigDecimal)
   implicit val arbJavaBigInteger = Arbitrary(genBigInt)
+
+  implicit val arbLocalTime = Arbitrary(genLocalTime)
+  implicit val arbLocalDate = Arbitrary(genLocalDate)
+  implicit val arbLocalDateTime = Arbitrary(genLocalDateTime)
+  implicit val arbMonthDay = Arbitrary(genMonthDay)
+  implicit val arbZoneOffset = Arbitrary(genZoneOffset)
+  implicit val arbOffsetDateTime = Arbitrary(genOffsetDateTime)
+  implicit val arbOffsetTime = Arbitrary(genOffsetTime)
+  implicit val arbYearMonth = Arbitrary(genYearMonth)
+  implicit val arbZoneId = Arbitrary(genZoneId)
+  implicit val arbZonedDateTime = Arbitrary(genZonedDateTime)
 }
