+++ tests/src/main/scala/pureconfig/data/Percentage.scala
