+++ tests/src/main/scala/pureconfig/equality/package.scala
